% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KneeArrow.R
\name{findCutoffCurvature}
\alias{findCutoffCurvature}
\title{Finds the point on the curve that has the maximum curvature}
\usage{
findCutoffCurvature(x, y)
}
\arguments{
\item{x}{x coordinates of points around the curve}

\item{y}{y coordinates of points around the curve}
}
\value{
(x, y) coordinates of the point with the greatest curvature
}
\description{
Finds the point on the curve that has the maximum curvature
}
