% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KneeArrow.R
\name{findCutoffFirstDerivative}
\alias{findCutoffFirstDerivative}
\title{Finds the point where the derivative is a fraction of the steepest slope}
\usage{
findCutoffFirstDerivative(x, y, slope_ratio = 0.5)
}
\arguments{
\item{x}{x coordinates of points around the curve}

\item{y}{y coordinates of points around the curve}

\item{slope_ratio}{the fraction of the steepest slope that defines knee point}
}
\value{
(x, y) coordinates of the knee point
}
\description{
Finds the point where the derivative is a fraction of the steepest slope
}
