% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KneeArrow.R
\name{findInverse}
\alias{findInverse}
\title{Finds value of x whose y value is y0}
\usage{
findInverse(x, y, y0)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{y0}{value for y for which we want corresponding x value}
}
\value{
value of the x coordinate whose y value is y0.
    If such x isn't in the domain of the function, then this function will return NA.
}
\description{
Finds value of x whose y value is y0
}
