% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KneeArrow.R
\name{inverse}
\alias{inverse}
\title{Inverse of a function}
\usage{
inverse(f, domain)
}
\arguments{
\item{f}{univariate function}

\item{domain}{domain of f given as (min, max) interval}
}
\value{
a function g such that f(x) = y and g(y) = x
}
\description{
Inverse of a function
}
\examples{
expinv <- inverse(exp, c(0,3))
expinv(exp(1))
}
